Implemented default values for function inputs.
This is one possible solution to issue #10. Right now default values are mandatory for all function inputs. If we made them optional we could either have a default "default value" (zero or something) or make it a compilation error for an input to have neither a default value nor an incoming edge.

For now I'm fine with defaults being mandatory, but I was thinking about how we would handle default values for dynamic inputs (e.g. a Summation function which users can add arbitrary numbers of inputs to and the output is the sum of all inputs). I think that would work fine with this impelmentation, just something to consider.